simplified interface on communicationAdapter to get ready for being typed, but leave the actual typing for later
