+++ src/services/platform/communication-adapter/communication-adapter.module.ts
@@ -7,7 +7,6 @@ import { MatrixAgentModule } from '../matrix/agent/matrix.agent.module';
 import { MatrixUserAdapterModule } from '../matrix/adapter-user/matrix.user.adapter.module';
 import { CommunicationAdapter } from './communication.adapter';
 import { SubscriptionModule } from '../subscription/subscription.module';
-import { MatrixMessageAdapterModule } from '../matrix/adapter-message/matrix.message.adapter.module';
 
 @Module({
   imports: [
@@ -15,7 +14,6 @@ import { MatrixMessageAdapterModule } from '../matrix/adapter-message/matrix.mes
     MatrixUserAdapterModule,
     MatrixRoomAdapterModule,
     MatrixGroupAdapterModule,
-    MatrixMessageAdapterModule,
     MatrixAgentModule,
     MatrixAgentPoolModule,
     SubscriptionModule,
