+++ src/services/platform/communication-adapter/dto/communication.dto.message.delete.ts
@@ -1,4 +1,4 @@
-export class CommunicationDeleteMessageFromCommunityRoomInput {
+export class CommunicationDeleteMessageInput {
   senderCommunicationsID!: string;
   messageId!: string;
   roomID!: string;
