--- src/services/platform/communication-adapter/dto/communication.dto.message.edit.community.ts
@@ -1,6 +1,0 @@
-export class CommunicationEditMessageOnCommunityRoomInput {
-  senderCommunicationsID!: string;
-  message!: string;
-  messageId!: string;
-  roomID!: string;
-}
