+++ src/services/platform/communication-adapter/dto/communication.dto.message.edit.ts
@@ -1,5 +1,6 @@
-export class CommunicationSendMessageCommunityInput {
+export class CommunicationEditMessageInput {
   senderCommunicationsID!: string;
   message!: string;
+  messageId!: string;
   roomID!: string;
 }
