+++ src/services/platform/communication-adapter/dto/communication.dto.message.send.ts
@@ -1,0 +1,5 @@
+export class CommunicationSendMessageInput {
+  senderCommunicationsID!: string;
+  message!: string;
+  roomID!: string;
+}
