+++ src/services/platform/matrix/adapter-room/matrix.room.adapter.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { MatrixMessageAdapterModule } from '../adapter-message/matrix.message.adapter.module';
 import { MatrixRoomAdapter } from './matrix.room.adapter';
 
 @Module({
-  imports: [],
+  imports: [MatrixMessageAdapterModule],
   providers: [MatrixRoomAdapter],
   exports: [MatrixRoomAdapter],
 })
